requirements.txt
# SPDX-FileCopyrightText: 2020 Melissa LeBlanc-Williams for Adafruit Industries
#
# SPDX-License-Identifier: Unlicense

Adafruit-Blinka
pillow
numpy
recordclass
